requirements.txt
selenium
chromedriver
python-wordpress-xmlrpc
